requirements.txt
  a n y i o = = 4 . 9 . 0   a t t r s = = 2 5 . 3 . 0   c e r t i f i = = 2 0 2 5 . 4 . 2 6   c f f i = = 1 . 1 7 . 1   c h a r s e t - n o r m a l i z e r = = 3 . 4 . 2 h11==0.16.0httpcore==1.0.9httpx==0.28.1idna==3.10numpy==2.2.5outcome==1.3.0.post0pandas==2.2.3pycparser==2.22PySocks==1.7.1python-dateutil==2.9.0.post0python-telegram-bot==22.1pytz==2025.2requests==2.32.3selenium==4.32.0six==1.17.0sniffio==1.3.1sortedcontainers==2.4.0ta==0.11.0trio==0.30.0trio-websocket==0.12.2typing_extensions==4.13.2tzdata==2025.2urllib3==2.4.0websocket-client==1.8.0wsproto==1.2.0  